+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to the "azure-search-analyze-client" extension will be docum
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [unrelease]
 
-- Initial release
+## [0.2.0] - 2019-03-17
+
+- Show original text in Response View
+- Show token start/end offset of the text in Response View
+- Show error message in Response View if there are errors in each requests
 
 ## [0.1.0] - 2019-03-13